Add reach-ui dialog and refactor board created dialog
